requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# LangChain and AI
langchain-nvidia-ai-endpoints>=0.0.5
langchain-groq>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20

# Data processing
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0

# Type checking and validation
pydantic>=2.0.0
typing-extensions>=4.8.0

# Google AI (if used)
google-genai>=0.2.0

# Image processing
Pillow>=10.0.0

# AutoViz for automated visualization
autoviz>=0.1.0

